+++ packages/uring_api/src/_uring_api/uring_api_capi_impl.h
@@ -33,6 +33,8 @@ int UringApiCapi_RingResetServing(PyObject *ring);
 int UringApiCapi_RingSubmit(PyObject *ring, int *submitted);
 int UringApiCapi_RingAutoSubmit(PyObject *ring, int *value);
 int UringApiCapi_RingSetAutoSubmit(PyObject *ring, int value);
+int UringApiCapi_RingWorkerAutoSubmit(PyObject *ring, int *value);
+int UringApiCapi_RingSetWorkerAutoSubmit(PyObject *ring, int value);
 int UringApiCapi_CompletionCheck(PyObject *completion);
 PyObject *UringApiCapi_CompletionUserData(PyObject *completion);
 int UringApiCapi_CompletionSetUserData(PyObject *completion, PyObject *value);